// sanity.types.ts
export type SanityImageDimensions = {
  _type: "sanity.imageDimensions";
  height?: number;
  width?: number;
  aspectRatio?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: ./src/sanity/lib/queries.ts
// Variable: POSTS_QUERY
// Query: *[_type == "post" && defined(slug.current)]{    _id,    body,    slug,    title,    categories[]->{      slug,      title    },    publishedAt,    _createdAt,    mainImage{      asset->{        url,        metadata{          dimensions,          lqip        }      }    }  }
export type POSTS_QUERYResult = Array<{
  _id: string;
  body: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?: "blockquote" | "h1" | "h2" | "h3" | "h4" | "normal";
        listItem?: "bullet";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  > | null;
  slug: Slug | null;
  title: string | null;
  categories: Array<{
    slug: Slug | null;
    title: string | null;
  }> | null;
  publishedAt: string | null;
  _createdAt: string;
  mainImage: {
    asset: {
      url: string | null;
      metadata: {
        dimensions: SanityImageDimensions | null;
        lqip: string | null;
      } | null;
    } | null;
  } | null;
}>;